requirements.txt
requests==2.32.5
openai==2.20.0
python-dotenv==1.2.1